new snap start/stop function for snap-1.0
Snap's latest version (v.1.0), comes with changes in the API and names of
binaries. Further, there's a now a JSON based task configuration file. In oder
to make use of the new snap 1.0 version, some changes to the initial setup has
to be done. The startup and tear down process has been changed to the snaptel
version, an template is added which contains all informations about the plugins
for the task.

Change-Id: Ieb4c8d16b46d03c424a1def8a4a233745b3d0d3d
Signed-off-by: Uwe Schilling <uwe.schilling@hlrs.de>